+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-
+- Fix error message prefix matching in IsAPINotAvailable().
 
 ## [0.2.2] 2020-04-06